Add set_user transaction callbacks
The set_user extension manipulates GUC state, and is therefore
responsible for maintaining it. In the event of transaction failure,
calls to the `(re)set_user` function were not correctly keeping track of
certain GUCs.

In an effort to make the extension handle transaction semantics more
robustly, introduce a transaction handler and add a formal state struct
to keep track of said state in a way that is (hopefully) clearer than
what was there before.